+++ src/aqua/graphics/__init__.py
@@ -25,7 +25,7 @@ from .multiple_maps import plot_maps, plot_maps_diff
 from .vertical_profiles import plot_vertical_profile, plot_vertical_profile_diff
 from .lat_lon_profiles import plot_lat_lon_profiles
 from .seasonal_means import plot_seasonal_lat_lon_profiles
-from .vertical_line import plot_vertical_line
+from .vertical_line import plot_vertical_lines
 
 
 __all__ = ["plot_gregory_monthly", "plot_gregory_annual",
@@ -38,7 +38,7 @@ __all__ = ["plot_gregory_monthly", "plot_gregory_annual",
            "plot_vertical_profile", "plot_vertical_profile_diff",
            "plot_lat_lon_profiles",
            "plot_seasonal_lat_lon_profiles",
-           "plot_vertical_line"]
+           "plot_vertical_lines"]
 
 
 def activate_style(style: str = None,