+++ src/aqua/graphics/vertical_line.py
@@ -1,26 +1,27 @@
 from typing import Optional, Tuple
 import matplotlib.pyplot as plt
-import numpy as np
 import xarray as xr
 from aqua.util import to_list
 from aqua.logger import log_configure
 from .styles import ConfigStyle
 
 
-def plot_vertical_line(data: xr.DataArray | list[xr.DataArray],
-                       ref_data: xr.DataArray = None,
-                       lev_name: str = "plev",
-                       lev_min: Optional[float] = None, lev_max: Optional[float] = None,
-                       labels: Optional[list[str]] = None,
-                       ref_label: Optional[str] = None,
-                       title: Optional[str] = None, title_size: Optional[int] = 16,
-                       style: Optional[str] = None,
-                       logscale: bool = False,
-                       invert_yaxis: bool = True,
-                       return_fig: bool = True, figsize: Tuple[int, int] = (10, 8),
-                       fig: Optional[plt.Figure] = None, ax: Optional[plt.Axes] = None,
-                       ax_pos: Tuple[int, int, int] = (1, 1, 1),
-                       loglevel: str = "WARNING"):
+def plot_vertical_lines(data: xr.DataArray | list[xr.DataArray],
+                        ref_data: xr.DataArray = None,
+                        lev_name: str = "plev",
+                        lev_min: Optional[float] = None, lev_max: Optional[float] = None,
+                        labels: Optional[list[str]] = None,
+                        ref_label: Optional[str] = None,
+                        label_size: Optional[int] = 14,
+                        axis_label_size: Optional[int] = 14,
+                        title: Optional[str] = None, title_size: Optional[int] = 18,
+                        style: Optional[str] = None,
+                        logscale: bool = False,
+                        invert_yaxis: bool = True,
+                        return_fig: bool = True, figsize: Tuple[int, int] = (6, 10),
+                        fig: Optional[plt.Figure] = None, ax: Optional[plt.Axes] = None,
+                        ax_pos: Tuple[int, int, int] = (1, 1, 1),
+                        loglevel: str = "WARNING"):
     """
     Plots a vertical line plot.
 
@@ -32,10 +33,13 @@ def plot_vertical_line(data: xr.DataArray | list[xr.DataArray],
         lev_max (float, optional): Maximum vertical level to plot.
         labels (list of str, optional): Labels for each DataArray.
         ref_label (str, optional): reference data label.
+        label_size (int, optional): Legend font size. Default is 14.
+        axis_label_size (int, optional): Axis label font size. Default is 14.
         title (str, optional): Plot title.
         title_size (int, optional): Title font size. Default is 16.
         style (str, optional): Plot style (default aqua style).
         logscale (bool, optional): Use log scale for y-axis if True.
+        invert_yaxis (bool, optional): Invert y-axis if True (default True).
         return_fig (bool, optional): If True, return (fig, ax).
         figsize (Tuple[int, int], optional): Figure size.
         fig (plt.Figure, optional): Optional figure to plot on.
@@ -74,8 +78,8 @@ def plot_vertical_line(data: xr.DataArray | list[xr.DataArray],
     xlabel = f"{var_name} ({var_units})" if var_name and var_units else "Unknown variable"
     ylabel = f"{lev_name} ({units})"  # Replace 'units' with actual units if available
 
-    ax.set_ylabel(ylabel)
-    ax.set_xlabel(xlabel)
+    ax.set_ylabel(ylabel, fontsize=axis_label_size)
+    ax.set_xlabel(xlabel, fontsize=axis_label_size)
 
     for i, d in enumerate(data):
         mask = (d[lev_name] >= lev_min) & (d[lev_name] <= lev_max)
@@ -88,7 +92,8 @@ def plot_vertical_line(data: xr.DataArray | list[xr.DataArray],
         mask = (ref_data[lev_name] >= lev_min) & (ref_data[lev_name] <= lev_max)
         ax.plot(ref_data.where(mask), ref_data[lev_name].where(mask), label=ref_label, linestyle='--', color='black')
 
-    ax.legend()
+    ax.legend(fontsize=label_size)
+    ax.grid(True, axis='y')
 
     if invert_yaxis:
         ax.invert_yaxis()