+++ CHANGELOG.md
@@ -14,7 +14,8 @@ Workflow modifications:
 
 AQUA core complete list:
 - Add updated grids conformal to OSI-SAF v3 (#2317)
-- Added graphical function for vertical profile plotting (#2314)
+- Area selection is now a separate class, `AreaSelection` in the `aqua.fldstat` module (#2245)
+- Added graphical function for vertical profile plotting (#2314, #2316)
 - Added catgen support for storyline experiments (#2308)
 - Pin maximum version of xarray (#2303)
 - CI/CD data now is read from aqua-dvc repository (#8370)
@@ -48,6 +49,7 @@ AQUA core complete list:
 - `apply_circular_window()` utility function to apply a circular window to cartopy maps (#2100)
 
 AQUA diagnostics complete list:
+- Timeseries: fix title and description generation (#2311)
 - Timeseries: update SeasonalCycles caption (#2299)
 - Seaice: added documentation and updated notebooks (#2249)
 - Seaice: update varname for PIOMAS and GIOMAS from sivol to sithick after updating the data in `obs` catalog (#2290)